Refactor api to own namespace modules
