+++ src/apis/healthcheck.py
@@ -1,0 +1,9 @@
+from flask_restx import Namespace, Resource
+
+api = Namespace('healthcheck')
+
+
+@api.route('/')
+class HealthCheck(Resource):
+    def get(self):
+        return {'status': 'ok'}
