+++ src/tests/test_api.py
@@ -20,7 +20,7 @@ class TestApiEndpoints(unittest.TestCase):
         targets = Target.objects.raw({
             '$or':
                 [{'source': {'$eq': 'ilmoitus'}},
-                {'name': {'$eq': 'overlapHylky'}}]
+                 {'name': {'$eq': 'overlapHylky'}}]
         }).all()
         for user in users:
             user.delete()
@@ -82,7 +82,8 @@ class TestApiEndpoints(unittest.TestCase):
                            )
         targetnote = Targetnote(diver=user, target=target)
         targetnote.save()
-        response = requests.get(f'{BASE_URL}/admin/pending?_end=10&_order=ASC&_sort=id&_start=0').json()
+        response = requests.get(
+            f'{BASE_URL}/admin/pending?_end=10&_order=ASC&_sort=id&_start=0').json()
         first_feature = response[0]
         self.assertGreater(len(response), 0)
         self.assertTrue(first_feature['is_pending'])
@@ -115,7 +116,8 @@ class TestApiEndpoints(unittest.TestCase):
                       is_ancient=False,
                       source='ilmoitus',
                       is_pending=True)
-        response = requests.get(f'{BASE_URL}/admin/duplicates?_end=10&_order=ASC&_sort=id&_start=0').json()
+        response = requests.get(
+            f'{BASE_URL}/admin/duplicates?_end=10&_order=ASC&_sort=id&_start=0').json()
         self.assertGreater(len(response), 0)
         self.assertIn(response[0]['id'], ['9999991', '99999992'])
         self.assertIn(response[1]['id'], ['9999991', '99999992'])
@@ -148,14 +150,15 @@ class TestApiEndpoints(unittest.TestCase):
                       is_ancient=False,
                       source='ilmoitus',
                       is_pending=False)
-        response = requests.get(f'{BASE_URL}/admin/duplicates?_end=10&_order=ASC&_sort=id&_start=0').json()
+        response = requests.get(
+            f'{BASE_URL}/admin/duplicates?_end=10&_order=ASC&_sort=id&_start=0').json()
         self.assertGreater(len(response), 0)
         self.assertIn(response[0]['id'], ['9999995', '99999994'])
         self.assertIn(response[1]['id'], ['9999995', '99999994'])
 
     def test_after_registering_login_with_correct_credentials_returns_auth(self):
         self.setUp()
-        res = requests.post(f'{BASE_URL}/register', json={
+        res = requests.post(f'{BASE_URL}/register/', json={
             'name': 'name',
             'email': 'email@email.com',
             'username': 'username4321',
@@ -163,7 +166,7 @@ class TestApiEndpoints(unittest.TestCase):
         })
         self.assertEqual(res.status_code, 200)
         self.assertNotIn('message', res.json().keys())
-        res = requests.post(f'{BASE_URL}/login', json={
+        res = requests.post(f'{BASE_URL}/login/', json={
             'username': 'username4321',
             'password': 'password'
         })
@@ -173,13 +176,13 @@ class TestApiEndpoints(unittest.TestCase):
 
     def test_token_required_route_requires_auth_token(self):
         self.setUp()
-        requests.post(f'{BASE_URL}/register', json={
+        requests.post(f'{BASE_URL}/register/', json={
             'name': 'name',
             'email': 'email@email.com',
             'username': 'username4321',
             'password': 'password'
         })
-        res = requests.post(f'{BASE_URL}/login', json={
+        res = requests.post(f'{BASE_URL}/login/', json={
             'username': 'username4321',
             'password': 'password'
         })
@@ -193,7 +196,7 @@ class TestApiEndpoints(unittest.TestCase):
 
     def test_after_correct_registering_login_with_incorrect_credentials_does_not_return_auth(self):
         self.setUp()
-        res = requests.post(f'{BASE_URL}/register', json={
+        res = requests.post(f'{BASE_URL}/register/', json={
             'name': 'name',
             'email': 'email@email.com',
             'username': 'username4321',
@@ -201,17 +204,18 @@ class TestApiEndpoints(unittest.TestCase):
         })
         self.assertEqual(res.status_code, 200)
         self.assertNotIn('message', res.json().keys())
-        res = requests.post(f'{BASE_URL}/login', json={
+        res = requests.post(f'{BASE_URL}/login/', json={
             'username': 'username4321',
             'password': 'pasword'
         })
         self.assertEqual(res.status_code, 200)
         self.assertNotIn('auth', res.json().keys())
-        self.assertEqual(res.json()['message'], 'Väärä käyttäjätunnus tai salasana')
+        self.assertEqual(res.json()['message'],
+                         'Väärä käyttäjätunnus tai salasana')
 
     def test_register_returns_error_message_if_username_is_taken(self):
         self.setUp()
-        res = requests.post(f'{BASE_URL}/register', json={
+        res = requests.post(f'{BASE_URL}/register/', json={
             'name': 'name',
             'email': 'email@email.com',
             'username': 'username4321',
@@ -219,7 +223,7 @@ class TestApiEndpoints(unittest.TestCase):
         })
         self.assertEqual(res.status_code, 200)
         self.assertNotIn('message', res.json().keys())
-        res = requests.post(f'{BASE_URL}/register', json={
+        res = requests.post(f'{BASE_URL}/register/', json={
             'name': 'name',
             'email': 'email@email.com',
             'username': 'username4321',
@@ -230,43 +234,43 @@ class TestApiEndpoints(unittest.TestCase):
 
     def test_user_targetnotes_returns_only_users_targetnotes(self):
         target1 = Target.create(target_id='9999999999991',
-                               name='Testihylky',
-                               town='SaimaaTesti',
-                               type='Hylky',
-                               x_coordinate=25.0,
-                               y_coordinate=61.0,
-                               location_method='gpstesti',
-                               location_accuracy='huonotesti',
-                               url='https://testiurl.com',
-                               created_at=datetime.datetime.now(),
-                               is_ancient=False,
-                               source='ilmoitus',
-                               is_pending=True)
+                                name='Testihylky',
+                                town='SaimaaTesti',
+                                type='Hylky',
+                                x_coordinate=25.0,
+                                y_coordinate=61.0,
+                                location_method='gpstesti',
+                                location_accuracy='huonotesti',
+                                url='https://testiurl.com',
+                                created_at=datetime.datetime.now(),
+                                is_ancient=False,
+                                source='ilmoitus',
+                                is_pending=True)
         target2 = Target.create(target_id='9999999999911',
-                               name='Testihylky2',
-                               town='SaimaaTesti2',
-                               type='Hylky',
-                               x_coordinate=25.0,
-                               y_coordinate=61.0,
-                               location_method='gpstesti',
-                               location_accuracy='huonotesti',
-                               url='https://testiurl.com',
-                               created_at=datetime.datetime.now(),
-                               is_ancient=False,
-                               source='ilmoitus',
-                               is_pending=True)
+                                name='Testihylky2',
+                                town='SaimaaTesti2',
+                                type='Hylky',
+                                x_coordinate=25.0,
+                                y_coordinate=61.0,
+                                location_method='gpstesti',
+                                location_accuracy='huonotesti',
+                                url='https://testiurl.com',
+                                created_at=datetime.datetime.now(),
+                                is_ancient=False,
+                                source='ilmoitus',
+                                is_pending=True)
         user1 = User.create(name='test user1',
-                           email='test1@example.com',
-                           phone='1234567',
-                           password='test',
-                           username='test1'
-                           )
+                            email='test1@example.com',
+                            phone='1234567',
+                            password='test',
+                            username='test1'
+                            )
         user2 = User.create(name='test user2',
-                           email='test2@example.com',
-                           phone='12345677',
-                           password='test',
-                           username='test2'
-                           )
+                            email='test2@example.com',
+                            phone='12345677',
+                            password='test',
+                            username='test2'
+                            )
         Targetnote.create(diver=user1, target=target1)
         Targetnote.create(diver=user2, target=target2)
         response = requests.get(f'{BASE_URL}/targets/user/test1').json()
@@ -274,23 +278,24 @@ class TestApiEndpoints(unittest.TestCase):
         first_feature = data[0]
         self.assertEqual(len(data), 1)
         self.assertTrue(first_feature['target']['properties']['is_pending'])
-        self.assertEqual(first_feature['target']['properties']['name'], 'Testihylky')
+        self.assertEqual(first_feature['target']
+                         ['properties']['name'], 'Testihylky')
         self.assertEqual(first_feature['diver']['name'], 'test user1')
 
     def test_user_targetnotes_returns_empty_data_dict_if_no_targetnotes(self):
         target2 = Target.create(target_id='9999999999911',
-                               name='Testihylky2',
-                               town='SaimaaTesti2',
-                               type='Hylky',
-                               x_coordinate=25.0,
-                               y_coordinate=61.0,
-                               location_method='gpstesti',
-                               location_accuracy='huonotesti',
-                               url='https://testiurl.com',
-                               created_at=datetime.datetime.now(),
-                               is_ancient=False,
-                               source='ilmoitus',
-                               is_pending=True)
+                                name='Testihylky2',
+                                town='SaimaaTesti2',
+                                type='Hylky',
+                                x_coordinate=25.0,
+                                y_coordinate=61.0,
+                                location_method='gpstesti',
+                                location_accuracy='huonotesti',
+                                url='https://testiurl.com',
+                                created_at=datetime.datetime.now(),
+                                is_ancient=False,
+                                source='ilmoitus',
+                                is_pending=True)
         User.create(name='test user1',
                     email='test1@example.com',
                     phone='1234567',
@@ -298,11 +303,11 @@ class TestApiEndpoints(unittest.TestCase):
                     username='test1'
                     )
         user2 = User.create(name='test user2',
-                           email='test2@example.com',
-                           phone='12345677',
-                           password='test',
-                           username='test2'
-                           )
+                            email='test2@example.com',
+                            phone='12345677',
+                            password='test',
+                            username='test2'
+                            )
         Targetnote.create(diver=user2, target=target2)
         response = requests.get(f'{BASE_URL}/targets/user/test1').json()
         data = response['data']
@@ -310,43 +315,43 @@ class TestApiEndpoints(unittest.TestCase):
 
     def test_user_dives_returns_only_users_dives(self):
         target1 = Target.create(target_id='9999999999991',
-                               name='Testihylky',
-                               town='SaimaaTesti',
-                               type='Hylky',
-                               x_coordinate=25.0,
-                               y_coordinate=61.0,
-                               location_method='gpstesti',
-                               location_accuracy='huonotesti',
-                               url='https://testiurl.com',
-                               created_at=datetime.datetime.now(),
-                               is_ancient=False,
-                               source='ilmoitus',
-                               is_pending=True)
+                                name='Testihylky',
+                                town='SaimaaTesti',
+                                type='Hylky',
+                                x_coordinate=25.0,
+                                y_coordinate=61.0,
+                                location_method='gpstesti',
+                                location_accuracy='huonotesti',
+                                url='https://testiurl.com',
+                                created_at=datetime.datetime.now(),
+                                is_ancient=False,
+                                source='ilmoitus',
+                                is_pending=True)
         target2 = Target.create(target_id='9999999999911',
-                               name='Testihylky2',
-                               town='SaimaaTesti2',
-                               type='Hylky',
-                               x_coordinate=25.0,
-                               y_coordinate=61.0,
-                               location_method='gpstesti',
-                               location_accuracy='huonotesti',
-                               url='https://testiurl.com',
-                               created_at=datetime.datetime.now(),
-                               is_ancient=False,
-                               source='ilmoitus',
-                               is_pending=True)
+                                name='Testihylky2',
+                                town='SaimaaTesti2',
+                                type='Hylky',
+                                x_coordinate=25.0,
+                                y_coordinate=61.0,
+                                location_method='gpstesti',
+                                location_accuracy='huonotesti',
+                                url='https://testiurl.com',
+                                created_at=datetime.datetime.now(),
+                                is_ancient=False,
+                                source='ilmoitus',
+                                is_pending=True)
         user1 = User.create(name='test user1',
-                           email='test1@example.com',
-                           phone='1234567',
-                           password='test',
-                           username='test1'
-                           )
+                            email='test1@example.com',
+                            phone='1234567',
+                            password='test',
+                            username='test1'
+                            )
         user2 = User.create(name='test user2',
-                           email='test2@example.com',
-                           phone='12345677',
-                           password='test',
-                           username='test2'
-                           )
+                            email='test2@example.com',
+                            phone='12345677',
+                            password='test',
+                            username='test2'
+                            )
         Dive.create(diver=user1,
                     target=target1,
                     location_correct=True,
@@ -378,48 +383,49 @@ class TestApiEndpoints(unittest.TestCase):
         data = response['data']
         first_feature = data[0]
         self.assertEqual(len(data), 2)
-        self.assertIn(first_feature['target']['properties']['name'], ['Testihylky2', 'Testihylky'])
+        self.assertIn(first_feature['target']['properties']['name'], [
+                      'Testihylky2', 'Testihylky'])
         self.assertEqual(first_feature['diver']['name'], 'test user1')
 
     def test_user_dives_return_empty_dict_if_no_dives(self):
         target1 = Target.create(target_id='9999999999991',
-                               name='Testihylky',
-                               town='SaimaaTesti',
-                               type='Hylky',
-                               x_coordinate=25.0,
-                               y_coordinate=61.0,
-                               location_method='gpstesti',
-                               location_accuracy='huonotesti',
-                               url='https://testiurl.com',
-                               created_at=datetime.datetime.now(),
-                               is_ancient=False,
-                               source='ilmoitus',
-                               is_pending=True)
+                                name='Testihylky',
+                                town='SaimaaTesti',
+                                type='Hylky',
+                                x_coordinate=25.0,
+                                y_coordinate=61.0,
+                                location_method='gpstesti',
+                                location_accuracy='huonotesti',
+                                url='https://testiurl.com',
+                                created_at=datetime.datetime.now(),
+                                is_ancient=False,
+                                source='ilmoitus',
+                                is_pending=True)
         Target.create(target_id='9999999999911',
-                               name='Testihylky2',
-                               town='SaimaaTesti2',
-                               type='Hylky',
-                               x_coordinate=25.0,
-                               y_coordinate=61.0,
-                               location_method='gpstesti',
-                               location_accuracy='huonotesti',
-                               url='https://testiurl.com',
-                               created_at=datetime.datetime.now(),
-                               is_ancient=False,
-                               source='ilmoitus',
-                               is_pending=True)
+                      name='Testihylky2',
+                      town='SaimaaTesti2',
+                      type='Hylky',
+                      x_coordinate=25.0,
+                      y_coordinate=61.0,
+                      location_method='gpstesti',
+                      location_accuracy='huonotesti',
+                      url='https://testiurl.com',
+                      created_at=datetime.datetime.now(),
+                      is_ancient=False,
+                      source='ilmoitus',
+                      is_pending=True)
         User.create(name='test user1',
-                           email='test1@example.com',
-                           phone='1234567',
-                           password='test',
-                           username='test1'
-                           )
+                    email='test1@example.com',
+                    phone='1234567',
+                    password='test',
+                    username='test1'
+                    )
         user2 = User.create(name='test user2',
-                           email='test2@example.com',
-                           phone='12345677',
-                           password='test',
-                           username='test2'
-                           )
+                            email='test2@example.com',
+                            phone='12345677',
+                            password='test',
+                            username='test2'
+                            )
         Dive.create(diver=user2,
                     target=target1,
                     location_correct=True,
